rename contrib modules, and optimize new_contrib_module.sh
