+++ app/src/main/java/com/github/piasy/bootstrap/di/AppComponent.java
@@ -31,8 +31,8 @@ import com.github.piasy.bootstrap.base.utils.UtilsModule;
 import com.github.piasy.bootstrap.features.profile.ProfileComponent;
 import com.github.piasy.bootstrap.features.search.SearchComponent;
 import com.github.piasy.bootstrap.features.splash.SplashComponent;
-import com.github.piasy.bootstrap.misc.MiscModule;
-import com.github.piasy.bootstrap.users.UsersModule;
+import com.github.piasy.octostars.misc.MiscModule;
+import com.github.piasy.octostars.users.UsersModule;
 import dagger.Component;
 import javax.inject.Singleton;
