+++ app/src/main/java/com/github/piasy/bootstrap/features/search/SearchUserResultAdapter.java
@@ -35,8 +35,8 @@ import android.view.ViewGroup;
 import butterknife.ButterKnife;
 import com.facebook.drawee.view.SimpleDraweeView;
 import com.github.piasy.bootstrap.R;
-import com.github.piasy.bootstrap.users.GitHubUser;
-import com.github.piasy.bootstrap.users.UserRepo;
+import com.github.piasy.octostars.users.GitHubUser;
+import com.github.piasy.octostars.users.UserRepo;
 import com.joanzapata.iconify.widget.IconTextView;
 import java.util.ArrayList;
 import java.util.List;
