+++ buildsystem/checkcC.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
 adb shell setprop dalvik.vm.dexopt-flags v=n,o=v && \
-./gradlew \
-:contrib:piasy:users:cC \
-:app:cC --stacktrace
+./gradlew --stacktrace :app:cC \
+:contrib:piasy:users:cC
