+++ buildsystem/ci.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-./gradlew clean :base:check :base:jacocoReport \
+./gradlew --stacktrace clean :base:check :base:jacocoReport \
+:app:check :app:jacocoReport \
 :contrib:piasy:misc:check :contrib:piasy:misc:jacocoReport \
-:contrib:piasy:users:check :contrib:piasy:users:jacocoReport \
-:app:check :app:jacocoReport --stacktrace
+:contrib:piasy:users:check :contrib:piasy:users:jacocoReport
