+++ contrib/piasy/misc/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="com.github.piasy.bootstrap.model"
+        package="com.github.piasy.octostars.misc"
         >
 
     <uses-permission android:name="android.permission.INTERNET"/>
