+++ contrib/piasy/users/src/main/AndroidManifest.xml
@@ -22,4 +22,4 @@
   ~ SOFTWARE.
   -->
 
-<manifest package="com.github.piasy.bootstrap.users"/>
+<manifest package="com.github.piasy.octostars.users"/>
