+++ contrib/piasy/users/src/main/java/com/github/piasy/octostars/users/UsersModule.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.github.piasy.bootstrap.users;
+package com.github.piasy.octostars.users;
 
 import com.google.gson.TypeAdapterFactory;
 import dagger.Module;
